New_testing_framework/requirements.txt
pytest~=8.2.0
allure-pytest #==2.13.5
httpx #==0.27.0
pytest_check #==2.3.1
playwright~=1.45
pytest-dotenv #==0.5.2
pydantic~=2.7.1
faker~=24.9.0
pillow~=10.3.0
pixelmatch~=0.3.0
typing_extensions~=4.11.0
jsonschema~=4.23.0
pytest-xdist~=3.6.1
